sprint-1/step-1 Исправление замечаний.
