+++ src/components/burger-constructor/burger-constructor.jsx
@@ -6,6 +6,7 @@ import {
 } from '@ya.praktikum/react-developer-burger-ui-components';
 import style from './burger-ingredients.module.css';
 import PropTypes from 'prop-types';
+import { ingredientsItemType } from '../../types/prop-types';
 
 const BurgerConstructor = ({ mainItem, bunItem }) => {
   return (
@@ -76,21 +77,6 @@ const BurgerConstructor = ({ mainItem, bunItem }) => {
     </div>
   );
 };
-
-const ingredientsItemType = PropTypes.shape({
-  calories: PropTypes.number,
-  carbohydrate: PropTypes.number,
-  fat: PropTypes.number,
-  image: PropTypes.string.isRequired,
-  image_large: PropTypes.string,
-  image_mobile: PropTypes.string,
-  name: PropTypes.string.isRequired,
-  price: PropTypes.number.isRequired,
-  proteins: PropTypes.number,
-  type: PropTypes.string.isRequired,
-  __v: PropTypes.number,
-  _id: PropTypes.string.isRequired,
-});
 
 BurgerConstructor.propTypes = {
   bunItem: PropTypes.arrayOf(ingredientsItemType),
