+++ src/components/burger-ingredients/ingredient-сard/ingredient-сard.module.css
@@ -1,0 +1,30 @@
+.card {
+  position: relative;
+  display: flex;
+  justify-content: end;
+}
+
+.counter {
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #4c4cff;
+  width: 32px;
+  height: 32px;
+  border-radius: 16px;
+}
+
+.cardPrice {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+}
+
+.cardTitle {
+  display: flex;
+  justify-content: center;
+  text-align: center;
+  max-width: 272px;
+}
